back_agroconnect/requirements.txt
fastapi==0.135.1
uvicorn==0.41.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pydantic==2.12.5
pydantic_core==2.41.5
python-multipart==0.0.12
cloudinary==1.38.0
anyio==4.12.1
starlette==0.52.1
typing_extensions==4.15.0
alembic==1.13.1
